refactor: insight package rename
