+++ package.json
@@ -9,38 +9,38 @@
     "prettier": " prettier -c \"**/*\"",
     "prettier:ca": "cd ./packages/chart-advisor && npm run prettier",
     "prettier:ckb": "cd ./packages/ckb && npm run prettier",
-    "prettier:insight": "cd ./packages/insight && npm run prettier",
+    "prettier:insight": "cd ./packages/lite-insight && npm run prettier",
     "prettier:sb": "cd ./packages/smart-board && npm run prettier",
     "prettier-fix": "npm run prettier -- --write",
     "prettier:packages": "run-p prettier:*",
     "format": "npm run prettier-fix && npm run lint-fix",
     "format:ca": "cd ./packages/chart-advisor && npm run format",
     "format:ckb": "cd ./packages/ckb && npm run format",
-    "format:insight": "cd ./packages/insight && npm run format",
+    "format:insight": "cd ./packages/lite-insight && npm run format",
     "format:sb": "cd ./packages/smart-board && npm run format",
     "format:packages": "lerna run format",
     "test": "jest",
     "test:ca": "cd ./packages/chart-advisor && npm run test",
     "test:ckb": "cd ./packages/ckb && npm run test",
-    "test:li": "cd ./packages/insight && npm run test",
+    "test:li": "cd ./packages/lite-insight && npm run test",
     "test:sb": "cd ./packages/smart-board && npm run test",
     "coverage": "jest --coverage",
     "test-live": "sh ./scripts/test-live",
     "size:ca": "cd ./packages/chart-advisor && limit-size",
     "size:ckb": "cd ./packages/ckb && limit-size",
-    "size:insight": "cd ./packages/insight && limit-size",
+    "size:insight": "cd ./packages/lite-insight && limit-size",
     "size:sb": "cd ./packages/smart-board && limit-size",
     "size:packages": "run-p size:*",
     "build:ca": "cd ./packages/chart-advisor && npm run build",
     "build:ckb": "cd ./packages/ckb && npm run build",
-    "build:insight": "cd ./packages/insight && npm run build",
+    "build:insight": "cd ./packages/lite-insight && npm run build",
     "build:sb": "cd ./packages/smart-board && npm run build",
     "build:dw": "cd ./packages/data-wizard && npm run build",
     "build:packages": "lerna run build",
     "build": "npm run build:packages",
     "clean:ca": "cd ./packages/chart-advisor && npm run clean",
     "clean:ckb": "cd ./packages/ckb && npm run clean",
-    "clean:insight": "cd ./packages/insight && npm run clean",
+    "clean:insight": "cd ./packages/lite-insight && npm run clean",
     "clean:sb": "cd ./packages/smart-board && npm run clean",
     "clean:packages": "lerna run clean",
     "clean:site": "cd ./packages/ava-site && npm run clean",
@@ -48,7 +48,7 @@
     "clear:ca": "cd ./packages/chart-advisor && rm -rf node_modules",
     "clear:ckb": "cd ./packages/ckb && rm -rf node_modules",
     "clear:demo": "cd ./demo && rm -rf node_modules",
-    "clear:insight": "cd ./packages/insight && rm -rf node_modules",
+    "clear:insight": "cd ./packages/lite-insight && rm -rf node_modules",
     "clear:sb": "cd ./packages/smart-board && rm -rf node_modules",
     "clear:packages": "lerna clean -y",
     "clear:root": "rimraf node_modules",
