+++ packages/lite-insight/.npmignore
