+++ packages/lite-insight/README.md
