+++ packages/lite-insight/jest.config.js
