+++ packages/lite-insight/package.json
