+++ packages/lite-insight/rollup.config.js
