+++ packages/lite-insight/src/algorithms/IQR.ts
