+++ packages/lite-insight/src/algorithms/base/compare.ts
