+++ packages/lite-insight/src/algorithms/index.ts
