+++ packages/lite-insight/src/algorithms/mannkendall-test.ts
