+++ packages/lite-insight/src/algorithms/pettitt-test.ts
