+++ packages/lite-insight/src/constant.ts
