+++ packages/lite-insight/src/insights/checkers.ts
