+++ packages/lite-insight/src/insights/extractors/categoryOutlier.ts
