+++ packages/lite-insight/src/insights/extractors/timeSeriesOutlier.ts
