+++ packages/lite-insight/src/insights/extractors/trend.ts
