+++ packages/lite-insight/src/insights/index.ts
