+++ packages/lite-insight/src/insights/util.ts
