+++ packages/lite-insight/src/interface.ts
