+++ packages/lite-insight/src/pipeline/index.ts
