+++ packages/lite-insight/src/pipeline/preprocess.ts
