+++ packages/lite-insight/src/pipeline/util.ts
