+++ packages/lite-insight/src/utils/aggregate.ts
