+++ packages/lite-insight/src/utils/common.ts
