+++ packages/lite-insight/src/visualization/description.ts
