+++ packages/lite-insight/tsconfig.json
